feat: refactor Sidebar import path and create Sidebar component
